+++ library/Zend/Locale/Math/PhpMath.php
@@ -192,8 +192,8 @@ class Zend_Locale_Math_PhpMath extends Zend_Locale_Math
         if (empty($op2)) {
             return NULL;
         }
-        $op1 = self::normalize($op1);
-        $op2 = self::normalize($op2);
+        $op1 = (int) self::normalize($op1);
+        $op2 = (int) self::normalize($op2);
 
         if ((int)$op2 === 0) {
             return NULL;
